feat(micro): The initial micro code + require tests
This is the first commit of the micro benchmarks test suite.  It tests both the
new module and cached module require operations.  The require new operation was
harder to test since I have to muck with the global Module cache object.

But just running this on my local machine shows a great difference
between cached and new require operations, which is what is expected.